chore(integ): apply compareStackTemplate to the rest
