+++ internal/pkg/deploy/cloudformation/stack/lb_web_service_integration_test.go
@@ -123,7 +123,7 @@ func TestLoadBalancedWebService_TemplateInteg(t *testing.T) {
 			expectedBytes := []byte(expected)
 			mExpected := make(map[interface{}]interface{})
 			require.NoError(t, yaml.Unmarshal(expectedBytes, mExpected))
-			require.Equal(t, mExpected, mActual)
+			compareStackTemplate(t, mExpected, mActual)
 		})
 
 		testName = fmt.Sprintf("Parameter values should render properly/%s", name)
