+++ internal/pkg/deploy/cloudformation/stack/scheduled_job_integration_test.go
@@ -6,21 +6,19 @@
 package stack_test
 
 import (
-	"github.com/aws/copilot-cli/internal/pkg/config"
 	"io/ioutil"
 	"path/filepath"
 	"testing"
 
-	"gopkg.in/yaml.v3"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/copilot-cli/internal/pkg/addon"
+	"github.com/aws/copilot-cli/internal/pkg/config"
 	"github.com/aws/copilot-cli/internal/pkg/deploy/cloudformation/stack"
-	"github.com/aws/copilot-cli/internal/pkg/workspace"
-
 	"github.com/aws/copilot-cli/internal/pkg/manifest"
+	"github.com/aws/copilot-cli/internal/pkg/workspace"
 	"github.com/stretchr/testify/require"
+	"gopkg.in/yaml.v3"
 )
 
 const (
@@ -81,7 +79,7 @@ func TestScheduledJob_Template(t *testing.T) {
 		mExpected := make(map[interface{}]interface{})
 		require.NoError(t, yaml.Unmarshal(expectedBytes, mExpected))
 		// Cut out zip file from EnvControllerAction
-		require.Equal(t, mExpected, mActual)
+		compareStackTemplate(t, mExpected, mActual)
 	})
 
 	t.Run("Parameter values should render properly", func(t *testing.T) {
