+++ internal/pkg/deploy/cloudformation/stack/windows_lb_web_service_integration_test.go
@@ -17,12 +17,10 @@ import (
 	"github.com/aws/copilot-cli/internal/pkg/addon"
 	"github.com/aws/copilot-cli/internal/pkg/config"
 	"github.com/aws/copilot-cli/internal/pkg/deploy/cloudformation/stack"
-	"github.com/aws/copilot-cli/internal/pkg/workspace"
-	"gopkg.in/yaml.v3"
-
 	"github.com/aws/copilot-cli/internal/pkg/manifest"
-
+	"github.com/aws/copilot-cli/internal/pkg/workspace"
 	"github.com/stretchr/testify/require"
+	"gopkg.in/yaml.v3"
 )
 
 const (
@@ -88,7 +86,7 @@ func TestWindowsLoadBalancedWebService_Template(t *testing.T) {
 		expectedBytes := []byte(expected)
 		mExpected := make(map[interface{}]interface{})
 		require.NoError(t, yaml.Unmarshal(expectedBytes, mExpected))
-		require.Equal(t, mExpected, mActual)
+		compareStackTemplate(t, mExpected, mActual)
 	})
 
 	t.Run("Parameter values should render properly", func(t *testing.T) {
